refactor(tests): split cross-project smoke test into focused modules
Split tests/vibe3/integration/test_cross_project_prompt_smoke.py (599 lines)
into two focused test modules for better maintainability:

- test_asset_discovery.py (184 lines): Layer 1 asset discovery tests
- test_cli_subprocess.py (293 lines): Layer 2 CLI subprocess tests

Updated conftest.py to include shared fixtures for cross-project testing.

All 10 tests pass successfully.

Closes #1928
